# Remove recursive functions