This pull request resolves #8 by rewriting all recursive functions as loops.
#### Benchmark before and after:

| Methode Name | String Lengths | Time Before | Time After |
| --- | --- | --- | --- |
| longest_common_subsequence() | 20000 chars | 2354μs | 4488μs |
| longest_common_subsequences() | 20 chars | 6145ms | 1691ms |
| diff() | 4000 chars | 1030μs | 389μs |

There is a decent performance degradation for `longest_common_subsequence()`. This is probably since we reverse the vector at end insted of inserting into the vector in the right order. I going to experiment with different ways to fill the vector.

This request includes and uses #7. So this should be merge after #7 if excepted.