@henninglive 2016-10-02
I have tried a few different insertion techniques, but I am unable to recover the lost performance. I even tried the unsafe approach with set_len() and get_unchecked_mut(), but I saw no difference in performance. There is probably something else going on, but I don’t know if it is worth investigating. The time spent finding the sequence is pretty much insignificant compared to the time it takes to build the LCStable, which takes 2590ms for two string with 20 000 characters. When calculated this LCStable takes 3.1GB of memory.